Separating styles into different files
