+++ app/user/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { View, StyleSheet, Text } from 'react-native';
+import { StyleSheet, Text, View } from 'react-native';
 import AdBanner from '@components/banners/AdBanner';
 
 export function MainIndexUser() {
